invite + qr code reader (png file)
now, by clicking on invite, the qr code associated is the one generated by the request out-of-band/create-invitation. to "read" the qr code from an user to another, the user has to type the name of the other. For example, if tanguy wants to send a message to yann, yann has to generate the qr code (send invitation button), QRCodeyann.png is created. tanguy writes "yann" below the line "Enter the QR Code of the contact", and now he gets the right invitation